# llm-service/docs/contracts/llm-service.openapi.yaml
openapi: 3.1.0
info:
  title: LLM Service HTTP API
  version: 2.0.0
  license: {name: UNLICENSED, identifier: UNLICENSED}
  description: Contrato HTTP canónico de la implementación actual. Evalúa uso de IA; no expone corrección académica ni soluciones esperadas.
servers:
  - url: /api/llm
security:
  - serviceJwt: []
paths:
...
  /courses/{courseId}/rubrics/{rubricVersionId}:
    get:
      summary: Get Rubric Version
      operationId: getRubricVersion
      parameters: [{$ref: '#/components/parameters/CourseId'}, {$ref: '#/components/parameters/RubricVersionId'}]
      responses: {'200': {$ref: '#/components/responses/RubricVersion'}, '403': {$ref: '#/components/responses/Problem'}}
    patch:
      summary: Update Rubric Draft
      operationId: updateRubricDraft
      parameters: [{$ref: '#/components/parameters/CourseId'}, {$ref: '#/components/parameters/RubricVersionId'}, {$ref: '#/components/parameters/IfMatch'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {$ref: '#/components/requestBodies/RubricInput'}
      responses: {'200': {$ref: '#/components/responses/RubricVersion'}, '409': {$ref: '#/components/responses/Problem'}}
  /courses/{courseId}/rubrics/{rubricVersionId}/publish:
    post:
      summary: Publish Rubric Version
      operationId: publishRubricVersion
      parameters: [{$ref: '#/components/parameters/CourseId'}, {$ref: '#/components/parameters/RubricVersionId'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      responses: {'200': {$ref: '#/components/responses/RubricVersion'}, '422': {$ref: '#/components/responses/Problem'}}
...
components:
  securitySchemes:
    serviceJwt: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, format: uuid}}
    IfMatch: {name: If-Match, in: header, required: true, schema: {type: integer, minimum: 1}}
    CourseId: {name: courseId, in: path, required: true, schema: {type: string, format: uuid}}
    RubricVersionId: {name: rubricVersionId, in: path, required: true, schema: {type: string, format: uuid}}
...
  requestBodies:
    RubricInput: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/RubricInput'}}}}
...
  responses:
    Problem: {description: RFC 7807 Problem Details}
    RubricVersion: {description: Rúbrica, content: {application/json: {schema: {$ref: '#/components/schemas/RubricVersion'}}}}
...
  schemas:
...
    RubricAnchor:
      type: object
      required: [behavior, referenceScore, example]
      properties: {behavior: {type: string}, referenceScore: {type: integer, minimum: 0, maximum: 100}, example: {type: string}}
    RubricAnchors:
      type: object
      required: [low, medium, high]
      properties: {low: {$ref: '#/components/schemas/RubricAnchor'}, medium: {$ref: '#/components/schemas/RubricAnchor'}, high: {$ref: '#/components/schemas/RubricAnchor'}}
    RubricDimension:
      type: object
      required: [key, criterion, anchors, weight]
      properties: {key: {type: string, enum: [AUTONOMY, CLARITY, PROGRESSION, COMPLIANCE, EFFICIENCY]}, label: {type: string}, criterion: {type: string}, anchors: {$ref: '#/components/schemas/RubricAnchors'}, weight: {type: number, exclusiveMinimum: 0, maximum: 100}}
    RubricInput:
      type: object
      required: [name, dimensions]
      properties: {name: {type: string}, dimensions: {type: array, minItems: 5, maxItems: 5, items: {$ref: '#/components/schemas/RubricDimension'}}}
    RubricVersion:
      allOf: [{$ref: '#/components/schemas/RubricInput'}, {type: object, required: [id, familyId, version, state, revision], properties: {id: {type: string, format: uuid}, familyId: {type: string, format: uuid}, version: {type: integer}, state: {type: string, enum: [DRAFT, PUBLISHED, SUPERSEDED]}, revision: {type: integer}}}]
...